+++ src/net/imshit/element/bullet/EnemyBullet.java
@@ -5,7 +5,7 @@ package net.imshit.element.bullet;
  */
 public class EnemyBullet extends AbstractBullet {
 
-    public EnemyBullet(int locationX, int locationY, int speedX, int speedY, int power) {
+    public EnemyBullet(float locationX, float locationY, float speedX, float speedY, int power) {
         super(locationX, locationY, speedX, speedY, power);
     }
