+++ src/net/imshit/element/prop/factory/BulletPropFactory.java
@@ -7,7 +7,7 @@ import net.imshit.element.prop.BulletProp;
  */
 public class BulletPropFactory extends AbstractPropFactory {
     @Override
-    public BulletProp createProp(int locationX, int locationY) {
+    public BulletProp createProp(float locationX, float locationY) {
         return new BulletProp(locationX, locationY, speedX, speedY);
     }
 }
